Move mocks to test files to prevent Sonar penalty
